+++ CHANGELOG.md
@@ -5,6 +5,13 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [unreleased]
+### Add
+- ```amend``` command
+
+### Changed
+- Work item tags are now sorted alphabetically
+
+## [0.2.7]
 ### Fix
 - ```list``` command does not throw error if no tasks have been recorder yet
 - ```list``` command now shows the last state of the task and not the oldest one